# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GOtome API",
    "version": "1",
    "description": "Every failure is answered with the ErrorEnvelope, whatever the status."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/auth/methods": {
      "get": {
        "operationId": "getSignInMethods",
        "summary": "How one may sign in: with a password, through an identity provider, or both",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInMethods"
                }
              }
            }
          },
          "default": {
            "description": "The request failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-permission": "public"
      }
    },
    "/auth/oidc/callback": {
      "get": {
        "operationId": "finishSingleSignOn",
        "summary": "Where the identity provider sends the browser back to; redirects into the app",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "The state the sign-in was started with",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "description": "The authorization code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iss",
            "in": "query",
            "description": "The issuer of the answer (RFC 9207)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "What the identity provider refused, instead of a code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "default": {
            "description": "The request failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-permission": "public"
      }
    },
...
    "/setup": {
      "get": {
        "operationId": "getSetup",
        "summary": "Whether the first account still has to be created",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupStatus"
                }
              }
            }
          },
          "default": {
            "description": "The request failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-permission": "public"
      },
      "post": {
        "operationId": "completeSetup",
        "summary": "Create the first account, an administrator, and sign in as it",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "The request failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-permission": "public"
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "type": "object",
        "description": "Why a request failed.",
        "properties": {
          "code": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SetupRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "SetupStatus": {
        "type": "object",
        "properties": {
          "needed": {
            "type": "boolean"
          }
        },
        "required": [
          "needed"
        ]
      },
      "SignInMethods": {
        "type": "object",
        "properties": {
          "password": {
            "type": "boolean"
          },
          "sso": {
            "$ref": "#/components/schemas/SsoMethod"
          }
        },
        "required": [
          "password"
        ]
      },
...
      "SsoMethod": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "editor",
              "reader"
            ]
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "role"
        ]
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gotome_session",
        "description": "The session cookie that signing in sets."
      }
    }
  }
}