# backend/internal/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"time"

	"github.com/google/uuid"
)

type AudioChapter struct {
	FileID   uuid.UUID
	Position int32
	Title    string
	StartMs  int64
	EndMs    int64
}

type Author struct {
	ID        uuid.UUID
	Name      string
	SortName  string
	NameKey   string
	CreatedAt time.Time
}

type Book struct {
	ID                 uuid.UUID
	LibraryID          uuid.UUID
	Title              string
	SortTitle          string
	TitleKey           string
	Subtitle           *string
	Description        *string
	Language           *string
	PublishedOn        *time.Time
	PublishedPrecision *string
	PublisherID        *uuid.UUID
	SeriesID           *uuid.UUID
	SeriesIndex        *float64
	ExternalRating     *float64
	PageCount          *int32
	CoverKey           *string
	LockedFields       []string
	FieldSources       []byte
	MergedIntoID       *uuid.UUID
	DeletedAt          *time.Time
	CreatedAt          time.Time
	UpdatedAt          time.Time
	PrimaryTextFileID  *uuid.UUID
	AuthorSort         string
	Placeholder        bool
}

type BookChunk struct {
	ID         int64
	BookID     uuid.UUID
	LibraryID  uuid.UUID
	FileID     uuid.UUID
	Position   int32
	Chapter    string
	PageFrom   *int32
	PageTo     *int32
	CharOffset int32
	Lang       string
	BodyEn     *string
	BodyDe     *string
	BodyXx     *string
}

type BookContributor struct {
	BookID   uuid.UUID
	AuthorID uuid.UUID
	Role     string
	Position int32
}

type BookFile struct {
	ID             uuid.UUID
	BookID         uuid.UUID
	LibraryID      uuid.UUID
	Kind           string
	Format         string
	RelPath        string
	SizeBytes      int64
	ModifiedAt     time.Time
	Sha256         []byte
	OriginalSha256 []byte
	ContentSha256  []byte
	PartIndex      *int32
	DurationMs     *int64
	PageCount      *int32
	PagesEstimated bool
	HasText        *bool
	Drm            bool
	ExtractState   string
	ExtractError   *string
	UploadedBy     *uuid.UUID
	MissingAt      *time.Time
	TrashedAt      *time.Time
	CreatedAt      time.Time
	UpdatedAt      time.Time
	TrackNumber    *int32
	DiscNumber     *int32
	ChunkedAt      *time.Time
	TrashedBy      *uuid.UUID
}

type BookIdentifier struct {
	ID     uuid.UUID
	BookID uuid.UUID
	FileID *uuid.UUID
	Type   string
	Value  string
}

type BookRelation struct {
	BookA uuid.UUID
	BookB uuid.UUID
	Kind  string
}

type BookTag struct {
	BookID uuid.UUID
	TagID  uuid.UUID
}

type BookVector struct {
	BookID       uuid.UUID
	Kind         string
	Model        string
	ModelVersion string
	SourceHash   []byte
	Embedding    string
	EmbeddedAt   time.Time
}

type BulkChange struct {
	ID         uuid.UUID
	CreatedBy  uuid.UUID
	SeesAll    bool
	Action     string
	Change     []byte
	CreatedAt  time.Time
	FinishedAt *time.Time
}

type BulkChangeBook struct {
	BulkChangeID uuid.UUID
	Position     int32
	BookID       uuid.UUID
	Outcome      *string
	Message      *string
	Skipped      []string
}

type Collection struct {
	ID          uuid.UUID
	OwnerID     uuid.UUID
	Name        string
	Description string
	Visibility  string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type CollectionItem struct {
	CollectionID uuid.UUID
	BookID       uuid.UUID
	Position     int32
	AddedAt      time.Time
}

type DuplicateEvidence struct {
	PairID uuid.UUID
	Kind   string
	Detail string
}

type DuplicatePair struct {
	ID        uuid.UUID
	BookA     uuid.UUID
	BookB     uuid.UUID
	State     string
	FoundAt   time.Time
	UpdatedAt time.Time
	Score     float32
}

type FileLsh struct {
	Band   int16
	Bucket int64
	FileID uuid.UUID
}

type FileSignature struct {
	FileID    uuid.UUID
	Version   int16
	Shingles  int32
	Signature []byte
	SignedAt  time.Time
}

type IndexVersion struct {
	Name      string
	Version   string
	UpdatedAt time.Time
}

type Library struct {
	ID         uuid.UUID
	Name       string
	RootPath   string
	Mode       string
	Writable   bool
	Visibility string
	OwnerID    *uuid.UUID
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type LibraryMember struct {
	LibraryID uuid.UUID
	UserID    uuid.UUID
	AddedAt   time.Time
}

type LibraryScan struct {
	ID            uuid.UUID
	LibraryID     uuid.UUID
	State         string
	RequestedBy   *uuid.UUID
	RequestedAt   time.Time
	JobID         *int64
	StartedAt     *time.Time
	FinishedAt    *time.Time
	FilesSeen     int32
	FilesAdded    int32
	FilesChanged  int32
	FilesMoved    int32
	FilesRestored int32
	FilesMissing  int32
	FilesSkipped  int32
	BooksAdded    int32
	Error         *string
}

type MetadataMatch struct {
	ID        uuid.UUID
	BookID    uuid.UUID
	Provider  string
	RecordID  string
	Score     float64
	Record    []byte
	State     string
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Notification struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Kind      string
	Data      []byte
	Link      string
	BookID    *uuid.UUID
	CreatedAt time.Time
	ReadAt    *time.Time
	LibraryID *uuid.UUID
}

type NotificationEvent struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	Kind      string
	LibraryID *uuid.UUID
	BookID    *uuid.UUID
	Data      []byte
	Link      string
	CreatedAt time.Time
}

type NotificationSubscription struct {
	UserID  uuid.UUID
	Kind    string
	Channel string
	Enabled bool
}

type OidcLogin struct {
	StateHash  []byte
	Nonce      string
	Verifier   string
	ReturnTo   string
	LinkUserID *uuid.UUID
	ExpiresAt  time.Time
}

type ProviderRecord struct {
	Provider   string
	RequestKey []byte
	Url        string
	Status     int32
	Body       []byte
	FetchedAt  time.Time
	ExpiresAt  time.Time
}

type Publisher struct {
	ID      uuid.UUID
	Name    string
	NameKey string
}

type ReadingFinish struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	BookID     uuid.UUID
	Medium     string
	FinishedAt time.Time
}

type ReadingProgress struct {
	UserID     uuid.UUID
	BookID     uuid.UUID
	Medium     string
	FileID     *uuid.UUID
	Locator    string
	Fraction   float64
	Chapter    *string
	Page       *int32
	PositionMs *int64
	ClientID   string
	UpdatedAt  time.Time
}

type ReadingSession struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	BookID       uuid.UUID
	Medium       string
	ClientID     string
	StartedAt    time.Time
	EndedAt      time.Time
	FromFraction float64
	ToFraction   float64
}

type SearchWord struct {
	Word string
}

type Series struct {
	ID        uuid.UUID
	Name      string
	NameKey   string
	CreatedAt time.Time
}

type Session struct {
	TokenHash  []byte
	UserID     uuid.UUID
	UserAgent  string
	CreatedAt  time.Time
	LastSeenAt time.Time
	ExpiresAt  time.Time
	ID         uuid.UUID
}

type Setting struct {
	Key       string
	Value     *string
	Sealed    []byte
	UpdatedAt time.Time
	UpdatedBy *uuid.UUID
}

type SmartShelf struct {
	ID         uuid.UUID
	OwnerID    uuid.UUID
	Name       string
	Filter     []byte
	Visibility string
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type StorageUse struct {
	UserID    *uuid.UUID
	UsedBytes int64
}

type Tag struct {
	ID      uuid.UUID
	Name    string
	NameKey string
}

type User struct {
	ID           uuid.UUID
	Username     string
	Email        *string
	PasswordHash *string
	Role         string
	QuotaBytes   *int64
	DisabledAt   *time.Time
	CreatedAt    time.Time
	UpdatedAt    time.Time
}

type UserBook struct {
	UserID     uuid.UUID
	BookID     uuid.UUID
	Status     string
	Rating     *int16
	StartedOn  *time.Time
	FinishedOn *time.Time
	UpdatedAt  time.Time
}

type UserIdentity struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	Issuer     string
	Subject    string
	Email      *string
	CreatedAt  time.Time
	LastUsedAt time.Time
}